Change 'labels' and silenced various warnings
Since 'labels' is obsolete as of Gnu Emacs 24.3, I have added a
compatibility layer so that labels is called on appropriate systems and
cl-labels (the replacement) is called on appropriate systems.  This
commit also silences a number of warnings around some deface and
defcustom calls.  I'm silencing them until I decide what I want to do
with them.

Conflicts:
	ecb-buffertab.el
	ecb-symboldef.el