Add COPYING file
Added a COPYING file as requested in Issue #7.  While the files in ECB
show clearly that it is licensed under GPLv2 or later and points to
where to get the GPLv2 license from Emacs or where to write and get a
copy, there’s no reason we shouldn’t keep a copy locally.